+++ CHANGELOG.md
@@ -23,17 +23,20 @@
 
 Released on FIXME: ??
 
-> 🏄 Summer update 2021🌴
+> 🍹 Summer update 2021 🍨
 
 - **Open any file** in explorer:
   - Open file with default program for file type with `<V>`
   - Open file with a specific program with `<W>`
+- **In-app release notes**
+  - Possibility to see the release note of the new available release whenever a new version is available
+  - Just press `<CTRL+R>` when a new version is available from the auth activity to read the release notes
 - Bugfix:
   - Fixed broken input cursor when typing UTF8 characters (tui-realm 0.3.2)
 - Dependencies:
   - Added `open 1.7.0`
   - Updated `textwrap` to `0.14.0`
-  - Updated `tui-realm` to `0.4.1`
+  - Updated `tui-realm` to `0.4.3`
 
 ## 0.5.1